+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This file is a history of the changes made to @idearium/cli.
 
 - Added `c mk delete`, `c mk hosts`, `c mk ip`, `c mk start`, `c mk stop` and `c mk restart` to control Minikube.
 - Added `c hosts add`, `c hosts get` and `c hosts remove` to help with easy hosts management.
-- Added `c project env ls`, `c project env get` and `c project env set`, `c project name`, `c project organisation`, `c project prefix` to help with project environment management.
+- Added `c project env ls`, `c project env get` and `c project env set`, `c project init`, `c project name`, `c project organisation`, `c project prefix` to help with project environment management.
 - Added a `c d build` command.
 - Added `c kc apply`, `c kc build`, `c kc clean`, `c kc context`, `c kc context get`, `c kc context set`, `c kc deploy` and `c kc stop` commands to help with kubectrl.